Move integration tests for taxel-cli
