+++ taxel-cli/Cargo.toml
@@ -9,7 +9,7 @@ path = "src/main.rs"
 name = "taxel"
 
 [features]
-integration-test = []
+external-test = []
 
 [dependencies]
 eric-sdk = { workspace = true }
