+++ taxel-cli/src/cmd/send.rs
@@ -58,61 +58,3 @@ pub fn send(matches: &ArgMatches) -> Result<(), anyhow::Error> {
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{app, cmd};
-
-    #[test]
-    #[cfg_attr(not(feature = "integration-test"), ignore)]
-    fn test_send() {
-        let args = vec![
-            cmd::BIN,
-            cmd::SEND,
-            "--xml-file",
-            "../test_data/taxonomy/v6.5/SteuerbilanzAutoverkaeufer_PersG.xml",
-            "--certificate-file",
-            "../test_data/test-certificate.pfx",
-            "--password",
-            "123456",
-        ];
-
-        let app = app();
-        let matches = app.get_matches_from(args);
-        let subcommand_matches = matches.subcommand_matches(cmd::SEND).unwrap();
-        println!("subcommand matches: {:#?}", subcommand_matches);
-
-        let res = send(subcommand_matches);
-        println!("res: {:#?}", res);
-
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    #[cfg_attr(not(feature = "integration-test"), ignore)]
-    fn test_send_and_print() {
-        let args = vec![
-            cmd::BIN,
-            cmd::SEND,
-            "--xml-file",
-            "../test_data/taxonomy/v6.5/SteuerbilanzAutoverkaeufer_PersG.xml",
-            "--certificate-file",
-            "../test_data/test-certificate.pfx",
-            "--password",
-            "123456",
-            "--print",
-            "ebilanz.pdf",
-        ];
-
-        let app = app();
-        let matches = app.get_matches_from(args);
-        let subcommand_matches = matches.subcommand_matches(cmd::SEND).unwrap();
-        println!("subcommand matches: {:#?}", subcommand_matches);
-
-        let res = send(subcommand_matches);
-        println!("res: {:#?}", res);
-
-        assert!(res.is_ok())
-    }
-}
