+++ taxel-cli/src/cmd/validate.rs
@@ -46,53 +46,3 @@ pub fn validate(matches: &ArgMatches) -> Result<(), anyhow::Error> {
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{app, cmd};
-
-    #[test]
-    #[cfg_attr(not(feature = "integration-test"), ignore)]
-    fn test_validate() {
-        let args = vec![
-            cmd::BIN,
-            cmd::VALIDATE,
-            "--xml-file",
-            "../test_data/taxonomy/v6.5/SteuerbilanzAutoverkaeufer_PersG.xml",
-        ];
-
-        let app = app();
-        let matches = app.get_matches_from(args);
-        let subcommand_matches = matches.subcommand_matches(cmd::VALIDATE).unwrap();
-        println!("subcommand matches: {:#?}", subcommand_matches);
-
-        let res = validate(subcommand_matches);
-
-        println!("res: {:#?}", res);
-        assert!(res.is_ok())
-    }
-
-    #[test]
-    #[cfg_attr(not(feature = "integration-test"), ignore)]
-    fn test_validate_and_print() {
-        let args = vec![
-            cmd::BIN,
-            cmd::VALIDATE,
-            "--xml-file",
-            "../test_data/taxonomy/v6.5/SteuerbilanzAutoverkaeufer_PersG.xml",
-            "--print",
-            "ebilanz.pdf",
-        ];
-
-        let app = app();
-        let matches = app.get_matches_from(args);
-        let subcommand_matches = matches.subcommand_matches(cmd::VALIDATE).unwrap();
-        println!("subcommand matches: {:#?}", subcommand_matches);
-
-        let res = validate(subcommand_matches);
-
-        println!("res: {:#?}", res);
-        assert!(res.is_ok())
-    }
-}
